Client-Server files
I haven't really changed much of the code. I currently cannot get it working on a mac, think it may just be the firewall blocking. If anyone with Windows could take a look and see if they can get it working, then that would be really helpful. :) Just need to specify your IP address when running the ChatClient.